price type changed

+++ src/domain/Home.java
@@ -2,7 +2,7 @@ package domain;
 
 public class Home extends Residence {
 
-    public Home(double price, int squareMeter, int roomNumber, int saloonNumber) {
+    public Home(int price, int squareMeter, int roomNumber, int saloonNumber) {
         super(price, squareMeter, roomNumber, saloonNumber);
     }
