+++ src/domain/Villa.java
@@ -1,7 +1,7 @@
 package domain;
 
 public class Villa extends Residence{
-    public Villa(double price, int squareMeter, int roomNumber, int saloonNumber) {
+    public Villa(int price, int squareMeter, int roomNumber, int saloonNumber) {
         super(price, squareMeter, roomNumber, saloonNumber);
     }
 }
